hotifx: long instead of int to avoid overflow exception

+++ src/Prtg.RabbitMq.Sensor/Program.cs
@@ -136,35 +136,35 @@ namespace Prtg.RabbitMq.Sensor
             {
                 channel = "Used Memory",
                 unit = PrtgUnit.BytesMemory,
-                value = ReadInt32(() => node.mem_used)
+                value = ReadInt64(() => node.mem_used)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Disk Free",
                 unit = PrtgUnit.BytesDisk,
-                value = ReadInt32(() => node.disk_free)
+                value = ReadInt64(() => node.disk_free)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Used File Descriptors",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => node.fd_used)
+                value = ReadInt64(() => node.fd_used)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Used Socket Descriptors",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => node.sockets_used)
+                value = ReadInt64(() => node.sockets_used)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Used Erlang Processes",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => node.proc_used)
+                value = ReadInt64(() => node.proc_used)
             });
 
             response.result.Add(new PrtgResult
@@ -185,21 +185,21 @@ namespace Prtg.RabbitMq.Sensor
             {
                 channel = "Used Memory per Second",
                 unit = PrtgUnit.BytesMemory,
-                value = ReadInt32(() => node.mem_used_details.rate)
+                value = ReadInt64(() => node.mem_used_details.rate)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "GC per Second",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => node.gc_num_details.rate)
+                value = ReadInt64(() => node.gc_num_details.rate)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "GC Bytes Reclaimed per Second",
                 unit = PrtgUnit.BytesMemory,
-                value = ReadInt32(() => node.gc_bytes_reclaimed_details.rate)
+                value = ReadInt64(() => node.gc_bytes_reclaimed_details.rate)
             });
 
             response.result.Add(new PrtgResult
@@ -311,21 +311,21 @@ namespace Prtg.RabbitMq.Sensor
             {
                 channel = "Total",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.queue_totals.messages)
+                value = ReadInt64(() => data.queue_totals.messages)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Ready",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.queue_totals.messages_ready)
+                value = ReadInt64(() => data.queue_totals.messages_ready)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Unacked",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.queue_totals.messages_unacknowledged)
+                value = ReadInt64(() => data.queue_totals.messages_unacknowledged)
             });
 
             // message_stats
@@ -368,35 +368,35 @@ namespace Prtg.RabbitMq.Sensor
             {
                 channel = "Consumers",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.object_totals.consumers)
+                value = ReadInt64(() => data.object_totals.consumers)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Queues",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.object_totals.queues)
+                value = ReadInt64(() => data.object_totals.queues)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Exchanges",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.object_totals.exchanges)
+                value = ReadInt64(() => data.object_totals.exchanges)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Connections",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.object_totals.connections)
+                value = ReadInt64(() => data.object_totals.connections)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Channels",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.object_totals.channels)
+                value = ReadInt64(() => data.object_totals.channels)
             });
 
             return response;
@@ -410,21 +410,21 @@ namespace Prtg.RabbitMq.Sensor
             {
                 channel = "Total",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.messages)
+                value = ReadInt64(() => data.messages)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Ready",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.messages_ready)
+                value = ReadInt64(() => data.messages_ready)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Unacked",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.messages_unacknowledged)
+                value = ReadInt64(() => data.messages_unacknowledged)
             });
 
             response.result.Add(new PrtgResult
@@ -455,14 +455,14 @@ namespace Prtg.RabbitMq.Sensor
             {
                 channel = "Consumers",
                 unit = PrtgUnit.Count,
-                value = ReadInt32(() => data.consumers)
+                value = ReadInt64(() => data.consumers)
             });
 
             response.result.Add(new PrtgResult
             {
                 channel = "Used Memory",
                 unit = PrtgUnit.BytesMemory,
-                value = ReadInt32(() => data.memory)
+                value = ReadInt64(() => data.memory)
             });
 
             return response;
@@ -480,7 +480,7 @@ namespace Prtg.RabbitMq.Sensor
             }
         }
 
-        private static string ReadInt32(Func<int> func)
+        private static string ReadInt64(Func<long> func)
         {
             try
             {
